Only install tools if they're required

+++ serversetup.sh
@@ -14,24 +14,26 @@ sudo apt-get -y install         \
     wget                        \
     git	 >> ${FULLLOG}
 
-echo " - Installing Ansible" | tee -a ${FULLLOG}
+if ! which ansible >/dev/null ; then
+    echo " - Installing Ansible" | tee -a ${FULLLOG}
+    sudo apt-add-repository ppa:ansible/ansible -y >> ${FULLLOG} 2>&1
+    sudo apt-get update >> ${FULLLOG} 2>&1
+    sudo apt-get install -y ansible >> ${FULLLOG} 2>&1
+fi
+if ! which terraform >/dev/null ; then
+    echo " - Install Terraform" | tee -a ${FULLLOG}
+    cd /tmp
+    wget -O terraform.zip https://releases.hashicorp.com/terraform/0.11.8/terraform_0.11.8_linux_amd64.zip >> ${FULLLOG} 2>&1
+    unzip terraform.zip >> ${FULLLOG}
+    sudo mv terraform /usr/local/bin/terraform
+    rm terraform.zip
+fi
 
-sudo apt-add-repository ppa:ansible/ansible -y >> ${FULLLOG} 2>&1
-sudo apt-get update >> ${FULLLOG} 2>&1
-sudo apt-get install -y ansible >> ${FULLLOG} 2>&1
-
-echo " - Install Terraform" | tee -a ${FULLLOG}
-
-cd /tmp
-wget -O terraform.zip https://releases.hashicorp.com/terraform/0.11.8/terraform_0.11.8_linux_amd64.zip >> ${FULLLOG} 2>&1
-unzip terraform.zip >> ${FULLLOG}
-sudo mv terraform /usr/local/bin/terraform
-rm terraform.zip
-
-echo " - Cloning Server Setup repository" | tee -a ${FULLLOG} 2>&1
-
-cd /tmp
-git clone https://github.com/epiclongstone/server-setup >> ${FULLLOG}
+if ! [ -d /tmp/server-setup ]; then
+    echo " - Cloning Server Setup repository" | tee -a ${FULLLOG} 2>&1
+    cd /tmp
+    git clone https://github.com/epiclongstone/server-setup >> ${FULLLOG}
+fi
 
 echo " - Running Ansible" | tee -a ${FULLLOG}
 cd /tmp/server-setup/ansible
@@ -46,18 +48,24 @@ if ! [ -s ${VAULTFILE} ]; then
     echo ${VAULTPASSWD} > ${VAULTFILE}
 fi
 
-echo " - Decrypting inventory file" | tee -a ${FULLLOG}
-ansible-vault decrypt inventory/longstone.yml --vault-password-file ${VAULTFILE}
 
-echo "Enter IP Address for longstone Server"
+echo -n "Enter IP Address for longstone Server: "
 read LONGSTONE_IP
+echo "Decrypting inventory file" | tee -a ${FULLLOG}
+ansible-vault decrypt inventory/longstone.yml --vault-password-file ${VAULTFILE}
 echo "Setting Longstone server IP to ${LONGSTONE_IP} in Ansible inventory" | tee -a ${FULLLOG}
 sed -i "s/{{ longstone_IP }}/${LONGSTONE_IP}/" inventory/longstone.yml
+echo "Encrypting inventory file" | tee -a ${FULLLOG}
+ansible-vault encrypt inventory/longstone.yml --vault-password-file ${VAULTFILE}
 
-echo " - Running Playbook" | tee -a ${FULLLOG}
+echo " - Running Playbook for all hosts" | tee -a ${FULLLOG}
 
-ansible-playbook -i inventory/longstone.yml ./all.yml --vault-password-file ~/.vault_pass | tee -a ${FULLLOG}
+#{
+ansible-playbook -i inventory/longstone.yml ./all.yml --vault-password-file ~/.vault_pass
 
-ansible-playbook -i inventory/longstone.yml ./rancher-server.yml --vault-password-file ~/.vault_pass | tee -a ${FULLLOG}
+echo " - Running Playbook for Rancher Server hosts" | tee -a ${FULLLOG}
+
+ansible-playbook -i inventory/longstone.yml ./rancher-server.yml --vault-password-file ~/.vault_pass
 
 
+#} | tee -a ${FULLLOG}
